Restructured tests subtree.
Details:
- Moved unit tests for zhmcclient module one level down into a 'zhmcclient'
  subdirectory.
- Moved unit tests in 'common' directory one level down into a 'tests/common'
  subdirectory.
- Moved the user example test case one level up into 'tests/unit'.
