+++ tests/unit/zhmcclient/test_adapter.py
@@ -24,7 +24,7 @@ import re
 
 from zhmcclient import Client, Adapter, NotFound, HTTPError
 from zhmcclient_mock import FakedSession
-from .utils import assert_resources
+from tests.common.utils import assert_resources
 
 
 # Object IDs and names of our faked adapters:
