+++ tests/unit/zhmcclient/test_cpc.py
@@ -24,7 +24,7 @@ import copy
 
 from zhmcclient import Client, Cpc, HTTPError
 from zhmcclient_mock import FakedSession
-from .utils import assert_resources
+from tests.common.utils import assert_resources
 
 
 # Object IDs and names of our faked CPCs:
