+++ tests/unit/zhmcclient/test_exceptions.py
