+++ tests/unit/zhmcclient/test_lpar.py
