+++ tests/unit/zhmcclient/test_manager.py
