+++ tests/unit/zhmcclient/test_partition.py
@@ -24,7 +24,7 @@ import copy
 
 from zhmcclient import Client, Partition, HTTPError, NotFound
 from zhmcclient_mock import FakedSession
-from .utils import assert_resources
+from tests.common.utils import assert_resources
 
 
 # Object IDs and names of our faked partitions:
