+++ tests/unit/zhmcclient/test_user.py
@@ -24,7 +24,7 @@ import copy
 
 from zhmcclient import Client, HTTPError, NotFound, User
 from zhmcclient_mock import FakedSession
-from .utils import assert_resources
+from tests.common.utils import assert_resources
 
 
 class TestUser(object):
