+++ tests/unit/zhmcclient/test_user_role.py
@@ -25,7 +25,7 @@ import six
 
 from zhmcclient import Client, HTTPError, NotFound, BaseResource, UserRole
 from zhmcclient_mock import FakedSession
-from .utils import assert_resources
+from tests.common.utils import assert_resources
 
 
 class TestUserRole(object):
