+++ tests/unit/zhmcclient/test_virtual_function.py
